+++ client/src/pages/Login.js
@@ -1,0 +1,17 @@
+import styled from 'styled-components';
+import LoginUl from '../components/login/LoginUl';
+
+function Login() {
+  return (
+    <ContainerDiv className="margininside">
+      <LoginUl />
+    </ContainerDiv>
+  );
+}
+
+export default Login;
+
+const ContainerDiv = styled.div`
+  min-height: 100vh;
+  justify-content: center;
+`;
